requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
channels==4.0.0
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==2.4.0
django-heroku==0.3.1
django-utils-six==2.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
redis==4.4.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
